Feat: Mapped the data from the payload for test
