+++ src/components/chartList/ChartList.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+
+
+function ChartList() {
+    
+  
+  return <div>
+      hello
+  </div>;
+}
+
+export default ChartList;
